move everything into derictories
