+++ src/app/app.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Page from './components/page';
+import Page from './components/page/page';
 
 const App = ({ url, data }) => (
   <Page url={url} title={data.canonical.headline} data={data} />
