+++ src/app/components/logo/logo.js
