+++ src/app/components/logo/logo.test.js
