+++ src/app/components/styled-icon-button/styled-icon-button.test.js
