+++ src/app/components/styled-share-bar/styled-share-bar.test.js
