+++ src/app/components/styled-time/styled-time.js
