+++ src/app/components/styled-time/styled-time.test.js
